RTCC code working.  Needs additional testing on molecules and cleanup.
