+++ pycc/tests/rtcc.py
@@ -12,9 +12,11 @@ import molecules as mol
 from lasers import sine_square_laser
 from lasers import gaussian_laser
 sys.path.insert(0, '../')
-import ode as myode
+#import ode as myode
 import numpy as np
-from scipy.integrate import ode
+from scipy.integrate import complex_ode as ode
+import time as timer
+from time_dep_lagranian import *
 
 # Psi4 Setup
 psi4.set_memory('2 GiB')
@@ -28,7 +30,7 @@ psi4.set_options({'basis': 'cc-pVDZ',
                   'd_convergence': 1e-13,
                   'r_convergence': 1e-13,
                   'diis': 1})
-mol = psi4.geometry(mol.moldict["He"])
+mol = psi4.geometry(mol.moldict["Be"])
 rhf_e, rhf_wfn = psi4.energy('SCF', return_wfn=True)
 
 ## Set up initial (t=0) amplitudes
@@ -47,39 +49,47 @@ ecc_test = ccdensity.compute_energy()
 print("ECC from density       = %20.15f" % ccdensity.ecc)
 print("ECC from wave function = %20.15f" % cc.ecc)
 
-# Gaussian pulse (a.u.)
-F_str = 1.0
-omega = 2.87
+print("Starting RTCC propagation...")
+time_init = timer.time()
+
+# Sine squared pulse (a.u.)
+F_str = 0.3
+omega = 0.2
 tprime = 5.0
 V = sine_square_laser(F_str, omega, tprime)
 
-axis = 2  # z-axis for field
 t0 = 0
-tf = 10.0
+tf = 1005.0
 h = 0.01
-rtcc = pycc.rtcc(cc, cclambda, ccdensity, V, axis)
+rtcc = pycc.rtcc(cc, cclambda, ccdensity, V)
 y0 = rtcc.collect_amps(cc.t1, cc.t2, cclambda.l1, cclambda.l2).astype('complex128')
-ODE = ode(rtcc.f).set_integrator('zvode')
+ODE = ode(rtcc.f).set_integrator('vode')
 ODE.set_initial_value(y0, t0)
 t = t0
 t1, t2, l1, l2 = rtcc.extract_amps(y0)
-dip0 = rtcc.dipole(t1, t2, l1, l2)
+mu_x, mu_y, mu_z = rtcc.dipole(t1, t2, l1, l2)
 ecc0 = rtcc.lagrangian(t, t1, t2, l1, l2)
 time = [t0]
-dip_z = [dip0]
-energy = [ecc0]
+dip_x = [mu_x]
+dip_y = [mu_y]
+dip_z = [mu_z]
+energy = [ecc0+rhf_e]
 print("Time(s)                  Energy (a.u.)                               Z-Dipole (a.u.)     ")
-print("%7.2f  %20.15f + %20.15fi  %20.15f + %20.15fi" % (t, ecc0.real+rhf_e, ecc0.imag, dip0.real, dip0.imag))
+print("%7.2f  %20.15f + %20.15fi  %20.15f + %20.15fi" % (t, ecc0.real+rhf_e, ecc0.imag, mu_z.real, mu_z.imag))
 
 while ODE.successful() and ODE.t < tf:
     y = ODE.integrate(ODE.t+h)
     t = ODE.t
     t1, t2, l1, l2 = rtcc.extract_amps(y)
-    dip = rtcc.dipole(t1, t2, l1, l2)
+    mu_x, mu_y, mu_z = rtcc.dipole(t1, t2, l1, l2)
     ecc = rtcc.lagrangian(t, t1, t2, l1, l2)
     time.append(t)
-    dip_z.append(dip)
+    dip_x.append(mu_x)
+    dip_y.append(mu_y)
+    dip_z.append(mu_z)
     energy.append(ecc+rhf_e)
-    print("%7.2f  %20.15f + %20.15fi  %20.15f + %20.15fi" % (t, ecc.real+rhf_e, ecc.imag, dip.real, dip.imag))
+    print("%7.2f  %20.15f + %20.15fi  %20.15f + %20.15fi" % (t, ecc.real+rhf_e, ecc.imag, mu_z.real, mu_z.imag))
 
-#np.savez("helium_cc-pvdz_F_str=10.0_omega=2.87.npz", time_points=time, energy=energy, dip_x=dip_z, dip_y=dip_z, dip_z=dip_z)
+np.savez("beryllium_cc-pvdz_F_str=0.3_omega=0.2.npz", time_points=time, energy=energy, dip_x=dip_x, dip_y=dip_y, dip_z=dip_z)
+
+print("\nRTCC propagation over %.2f a.u. completed in %.3f seconds.\n" % (tf-t0, timer.time() - time_init))
